+++ app.js
@@ -19,6 +19,8 @@ const path = require('path');
 const prom = require('./utils/prometheus');
 const swaggerSpec = require('./swagger.json');
 const swaggerUi = require('swagger-ui-express');
+const signoffController = require('./components/signoffs/signoffController');
+const authenticationController = require('./utils/apiAuth');
 
 const app = express();
 const port = process.env.PORT || 3000;
@@ -107,6 +109,12 @@ app.get('/metrics', async (req, res) => {
   res.set('Content-Type', prom.prometheus.register.contentType);
   res.end(await prom.prometheus.register.metrics());
 });
+
+app.post(`/case/:caseIdentifier/${urls.signoff}`, signoffController.addSignoff);
+app.post(`/case/${urls.signoff}`, signoffController.addBatchSignoffs);
+app.get(`/case/:caseIdentifier/${urls.signoff}`, signoffController.getSignoff);
+
+app.post('/token', authenticationController.addNewKey);
 
 app.use(errorHandler);
 app.use(prom.monitorAfterRequest);