+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project attempts to adhere to Semantic Versioning.
 
 ## UNRELEASED
 
+### Added
+
+* Sign-off records can be added at endpoint `/case/{caseIdentifier}/sign-off`
+* Sign-off records can be retrieved at endpoint `/case/{caseIdentifier}/sign-off`
+* New endpoints require api-key in header for authorization
+* Sign-off records can be bulk added at endpoint `/case/sign-off`
+
 ## 3.4.0 [2023-10-24]
 
 ### Added